% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{asuse_delete}
\alias{asuse_delete}
\title{Delete local ASUSE database}
\usage{
asuse_delete(ask = TRUE)
}
\arguments{
\item{ask}{whether to prompt the user for confirmation before deleting existing census databases. Defaults to TRUE.}
}
\value{
NULL (invisibly)
}
\description{
Removes the ASUSE directory and all its contents.
}
\examples{
\dontrun{ asuse_delete() }
}
