% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{asuse_file_path}
\alias{asuse_file_path}
\title{Local ASUSE database file path}
\usage{
asuse_file_path(dir = asuse_path())
}
\arguments{
\item{dir}{Path to the ASUSE directory on disk. By default this is #' "asuse" inside the user's R data directory, or the
directory specified by the \code{ASUSE_DIR} environment variable if set.}
}
\description{
Returns the path to the local ASUSE database directory.
}
\examples{
asuse_file_path()
}
