% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{plfs_delete}
\alias{plfs_delete}
\title{Delete local plfs database}
\usage{
plfs_delete(ask = TRUE)
}
\arguments{
\item{ask}{whether to prompt the user for confirmation before deleting existing census databases. Defaults to TRUE.}
}
\value{
NULL (invisibly)
}
\description{
Removes the plfs directory and all its contents.
}
\examples{
\dontrun{ plfs_delete() }
}
