% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_datasets.R
\docType{data}
\name{available_datasets}
\alias{available_datasets}
\title{Available datasets in the package}
\format{
A string vector
}
\usage{
available_datasets
}
\description{
Shows the available datasets in the package.
}
\examples{
available_datasets
}
\author{
Based on the Nesstar files from the Ministry of Statistics and Program Implementation
}
\keyword{data}
