% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plfs-package.R
\docType{package}
\name{plfs-package}
\alias{plfs}
\alias{plfs-package}
\title{plfs: Periodic Labour Force Survey (PLFS)}
\description{
Microdata from the Periodic Labour Force Survey (PLFS) conducted by the National Sample Survey Office (NSSO), Ministry of Statistics and Programme Implementation, Government of India. Tidied and organized using DuckDB for efficient querying.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
