% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plfs-package.R
\docType{package}
\name{plfs-package}
\alias{plfs}
\alias{plfs-package}
\title{plfs: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
