% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{plfs_file_path}
\alias{plfs_file_path}
\title{Local plfs database file path}
\usage{
plfs_file_path(dir = plfs_path())
}
\arguments{
\item{dir}{Path to the plfs directory on disk. By default this is #' "plfs" inside the user's R data directory, or the
directory specified by the \code{plfs_DIR} environment variable if set.}
}
\description{
Returns the path to the local plfs database directory.
}
\examples{
plfs_file_path()
}
